+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * (\#231) New utility for filtering ap in Astroutils
 * (\#238) New methods setOptionTimeMJD setOptionEnergybin
 * (\#242) tdms progress bars 
+* (\#252) New methods for setting dq
 * (anaconda package update) The installation time of Agilepy has been reduced a lot.
 
 ## Release 1.2.0 - 10/12/20